refactor(web): server fn のディレクトリを api/ に統一
- features/blog/server/* → features/blog/api/*
- features/og/server/* → features/og/api/*
- 既存の api/ 命名規約 (mutation 系) と統一
