+++ apps/web/src/features/blog/types.ts
@@ -1,5 +1,5 @@
-import type { getBlogPost } from "./server/get-blog-post";
-import type { getBlogPosts } from "./server/get-blog-posts";
+import type { getBlogPost } from "./api/get-blog-post";
+import type { getBlogPosts } from "./api/get-blog-posts";
 
 export type BlogPost = Awaited<ReturnType<typeof getBlogPost>>;
 export type BlogPostListItem = Awaited<ReturnType<typeof getBlogPosts>>[number];
